pyproject.toml
[project]
name = "pixelbot"
version = "1.1.0"
description = "Desktop automation helper for Path of Exile and other games"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.9.0",
    "simple-websocket>=1.0.0",
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "pyautogui>=0.9.54",
    "pydirectinput>=1.0.4",
    "pywebview>=5.0",
]

[project.scripts]
pixelbot = "pixelbot.__main__:main"

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pixelbot"]

[tool.uv]
package = true
